+++ src/Field/CloneField.php
@@ -247,4 +247,14 @@ class CloneField extends \Geniem\ACF\Field {
     public function get_clones() {
         return $this->clone;
     }
+
+    /**
+     * Clone fields do not support conditional logic
+     *
+     * @param mixed ...$args Any args this functions gets.
+     * @throws \Geniem\ACF\Exception The exception this always throws.
+     */
+    public function add_conditional_logic( \Geniem\ACF\ConditionalLogicGroup $group ) {
+        throw new \Geniem\ACF\Exception( 'Geniem\ACF\Field\Clone: add_conditional_logic() can\'t be called on this field type.' );
+    }
 }