+++ src/Field/File.php
@@ -57,7 +57,24 @@ class File extends \Geniem\ACF\Field {
      *
      * @var array
      */
-    protected $mime_types;
+    protected $mime_types = [];
+
+    /**
+     * Export field in ACF's native format.
+     *
+     * @param boolean $register Whether the field is to be registered.
+     *
+     * @return array
+     */
+    public function export( $register = false ) {
+        // Call the original export method
+        $obj = parent::export( $register );
+
+        // Convert the mime type array to a comma-separated list
+        $obj['mime_types'] = implode( ',', $obj['mime_types'] );
+
+        return $obj;
+    }
 
     /**
      * Set the return format of the field.