+++ src/Field/GroupableField.php
@@ -11,13 +11,20 @@ namespace Geniem\ACF\Field;
 abstract class GroupableField extends \Geniem\ACF\Field {
 
     /**
+     * Field variable
+     *
+     * @var string
+     */
+    protected $fields_var = 'sub_fields';
+
+    /**
      * Constructor.
      *
      * @param string      $label          Label for the field.
      * @param string|null $key            Key for the field.
      * @param string|null $name           Name for the field.
      */
-    public function __construct( string $label, string $key = null, string $name = null  ) {
+    public function __construct( string $label, string $key = null, string $name = null ) {
         // Add Groupable to its property to be pseudo-extended
         $this->groupable = new \Geniem\ACF\Field\Groupable( $this );
 
@@ -28,12 +35,70 @@ abstract class GroupableField extends \Geniem\ACF\Field {
     /**
      * Magic function __call
      *
-     * @param string $name       Function name to call
-     * @param array  $arguments  Function arguments
-     * @return mixed             Return value of the function
+     * @param string $name       Function name to call.
+     * @param array  $arguments  Function arguments.
+     * @return mixed             Return value of the function.
      */
     public function __call( $name, array $arguments ) {
         // Call the method from groupable
         return \call_user_func_array( [ $this->groupable, $name ], $arguments );
+    }
+
+    /**
+     * Clone method
+     *
+     * Forces the developer to give new key to cloned field.
+     *
+     * @param string $key  Field key.
+     * @param string $name Field name (optional).
+     * @return Geniem\ACF\Field
+     */
+    public function clone( $key, $name = null ) {
+        $clone = clone $this;
+
+        $clone->set_key( $key );
+
+        if ( isset( $name ) ) {
+            $clone->set_name( $name );
+        }
+
+        $field_map = [];
+
+        $clone->{ $this->fields_var } = array_map( function( $field ) use ( $key, &$field_map ) {
+            $clone = $field->clone( $key . '_' . $field->get_key() );
+
+            $field_map[] = [
+                'original' => $field,
+                'cloned'   => $clone,
+            ];
+
+            return $clone;
+
+        }, $clone->{ $this->fields_var });
+
+        if ( count( $field_map ) > 0 ) {
+            $clone->{ $this->fields_var } = array_map( function( $field ) use ( $field_map ) {
+                if ( count( $field->conditional_logic ) > 0 ) {
+                    foreach ( $field->conditional_logic as &$logics ) {
+                        if ( count( $logics ) > 0 ) {
+                            foreach ( $logics as &$logic ) {
+                                foreach ( $field_map as $map ) {
+                                    if ( $map['original'] === $logic['field'] ) {
+                                        $logic['field'] = $map['cloned'];
+                                        break;
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+
+                return $field;
+            }, $clone->{ $this->fields_var });
+        }
+
+        $clone->update_self( $clone );
+
+        return $clone;
     }
 }