+++ src/Field/Repeater.php
@@ -90,6 +90,20 @@ class Repeater extends \Geniem\ACF\Field\GroupableField {
             $obj['sub_fields'] = array_values( $obj['sub_fields'] );
         }
 
+        // Convert the wrapper class array to a space-separated string.
+        if ( isset( $obj['wrapper']['class'] ) && ! empty( $obj['wrapper']['class'] ) ) {
+            $obj['wrapper']['class'] = implode( ' ', $obj['wrapper']['class'] );
+        }
+        else {
+            $obj['wrapper']['class'] = '';
+        }
+
+        // Remove unnecessary properties from the exported array.
+        unset( $obj['inheritee'] );
+        unset( $obj['groupable'] );
+        unset( $obj['fields_var'] );
+        unset( $obj['filters'] );
+
         return $obj;
     }